gralloc3: [g19] allow disabling AFBC without warning logs
If AFBC is disabled for a format (e.g: rgb565),
but .bpp_afbc is still defined,
than gralloc throws a error message.

This patch allows .bpp_afbc to defined for a format
even if AFBC is disabled for that format.

This shouldn't be an error since even if a format supports AFBC,
the vendor or OEM should be able to disable it by setting
the .afbc field of format info to FALSE.
They shouldn't have to modify other attributes of the format.

Allowing .bpp_afbc to still be defined makes it easier
to enabling/disabling AFBC.

Change-Id: Ib931a49136dd49f75aeefc04a2a88ce998727f67
Signed-off-by: SeYeong Byeon <sy.byeon@samsung.com>